sstable: move obsoleteKey collector/filter code
